Rename dtcsimulation -> dtc_simulation
